requirements.txt
# requirements.txt
PyGitHub==1.55